+++ template/src/components/views/UserMenu.js
@@ -1,9 +1,7 @@
 import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-
 import { makeStyles } from '@material-ui/core/styles';
 import { Avatar, Button, Grid, Menu, MenuItem, Typography } from '@material-ui/core';
-
 import OAuthLogin from 'lib/sdk/oauth/OAuthLogin';
 import { logout } from 'config/oauthSlice';
