+++ template/src/components/views/datasets/Datasets.js
@@ -1,14 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { useSelector } from 'react-redux';
-
 import { makeStyles } from '@material-ui/core/styles';
 import { Grid, Typography } from '@material-ui/core';
-
-import { getDatasets } from 'lib/sdk';
-
+import { getUserDatasets } from 'lib/sdk';
 import { selectOAuthCredentials } from 'config/oauthSlice';
-
-import DatasetList from 'components/views/DatasetList';
+import UserDatasets from 'components/views/datasets/UserDatasets';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -31,23 +27,16 @@ const datasetsPagination = { page: 1, size: 50 };
 
 function Datasets() {
   const credentials = useSelector(selectOAuthCredentials);
-
   const classes = useStyles();
-
   const [datasets, setDatasets] = useState([]);
   const [loading, setLoading] = useState(false);
 
   useEffect(() => {
     if (credentials) {
-      // Get datasets, once logged in
+      // Get user datasets, once logged in
       setLoading(true);
-      getDatasets(credentials, datasetsPagination).then((data) => {
-        // just cartodbfied datasets can be loaded as deckgl layers with CartoSQLLayers...
-        const cartodbfied = data.result.filter(
-          (dataset) =>
-            dataset.cartodbfied && dataset.table_schema === credentials.username
-        );
-        setDatasets(cartodbfied);
+      getUserDatasets(credentials, datasetsPagination).then((datasets) => {
+        setDatasets(datasets);
         setLoading(false);
       });
     }
@@ -70,7 +59,7 @@ function Datasets() {
 
         {credentials ? (
           <Grid item className={classes.datasetList}>
-            <DatasetList datasets={datasets} loading={loading} />
+            <UserDatasets datasets={datasets} loading={loading} />
           </Grid>
         ) : (
           <Typography className={classes.datasetsNotAvailable}>
