+++ template/src/lib/sdk/index.js
@@ -3,3 +3,4 @@ export * from './widgets';
 export * from './utils';
 export * from './google';
 export * from './models';
+export * from './oauth';
