+++ template/src/lib/sdk/oauth/index.js
@@ -1,0 +1,7 @@
+export {
+  createOAuthPopup,
+  getOAuthParamsFromCallback,
+  buildTokenRenewalUrl,
+} from './oauthHelper';
+export { default as OAuthCallback } from './OAuthCallback';
+export { default as OAuthLogin } from './OAuthLogin';
