+++ template/src/routes.js
@@ -1,12 +1,11 @@
 import React from 'react';
+import { Navigate } from 'react-router-dom';
+import OAuthCallback from 'lib/sdk/oauth/OAuthCallback';
 import Home from 'components/views/Home';
 import Stores from 'components/views/stores/Stores';
 import StoresDetail from 'components/views/stores/StoresDetail';
-import { Navigate } from 'react-router-dom';
-
 import Kpi from 'components/views/Kpi';
-import Datasets from 'components/views/Datasets';
-import OAuthCallback from 'lib/sdk/oauth/OAuthCallback';
+import Datasets from 'components/views/datasets/Datasets';
 
 const routes = [
   {
